refactor: split ApiResource tests
