+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/test/java/io/gravitee/rest/api/management/v2/rest/resource/AbstractResourceTest.java
@@ -94,6 +94,12 @@ public abstract class AbstractResourceTest extends JerseySpringTest {
     @Autowired
     protected MediaService mediaService;
 
+    @Autowired
+    protected ParameterService parameterService;
+
+    @Autowired
+    protected WorkflowService workflowService;
+
     @Before
     public void setUp() {
         when(permissionService.hasPermission(any(), any(), any(), any())).thenReturn(true);
